+++ libbeat/idxmgmt/ilm/single.go
